refactor(core): make src/core module graph acyclic
Groundwork for splitting `openobserve-core` into per-domain crates so api/ and
jobs/ can depend on the services they use instead of one 70k-line crate. Rust
crates cannot be mutually recursive, so the reference graph inside core has to
be a DAG along the intended crate boundaries before anything can move out.

Every top-level module in src/core/src is now assigned a layer:

    sink < identity < alerting < ingest < dashboards < admin < top

and 16 references that pointed sideways or upwards are gone. The recurring fix
is splitting a module that mixed a dependency-free producer with a consumer
that pulls in a higher layer:

- cloud_events: the event types and `enqueue_cloud_event` (a queue push, no
  dependencies) move to `cloud_events`; the flusher, which needs the ingestion
  service, stays in `self_reporting`. Drops 7 refs from logs, traces and
  organization into the admin layer.
- workflows: definitions, CRUD and `send_workflow_trigger` stay in `workflows`
  so alerts can keep firing workflows as a notification target; everything that
  needs `ExecutablePipeline` to actually run one moves to `workflow_execution`.
- `self_reporting::search` and the `_o2_service_graph` reader only need the
  search service, so they become `usage_search` and `service_graph_query`.
  Incident topology enrichment no longer drags in traces.
- `organization::get_summary` aggregates streams, pipelines, alerts, functions
  and dashboards, so it moves up to `org_summary` rather than sitting in a
  module those same domains depend on.
- `map_error_to_http_response` is domain-agnostic and moves down to
  `http_error`. The `From<SomeError> for Response` impls stay in `http`: the
  orphan rule pins them to the crate declaring the error type, so they travel
  with those types when the domains are extracted.
- The pipeline-only `delete_backfill_jobs_by_pipeline` and the top-level
  `cache` module move next to what they serve.

scripts/check_core_layering.py enforces the layering and runs in CI. A module
added without a layer assignment fails the check, so the boundary stays a
deliberate choice. No behaviour change; this is all code movement.

`build_edges_sql`, added to `service_graph/api.rs` after this work started,
moves to `service_graph_query.rs` with its four tests, since it is only used by
the query that moved. The Cargo.lock line belongs to the dependency removal in
the preceding commit.
